CDAP-14022 add structure for bootstrap framework
Added a skeleton of the bootstrap framework.
Includes the classes that represent a bootstrap config and the steps
in the config. Also includes a skeleton of the BootstrapService
that will be in charge of loading the bootstrap config and executing
the steps.

Modified the build to contain a bootstrap file.