CDAP-14308 fix flaky SparkTest
The test could fail because of incorrect assumptions about
ProgramManager's waitForRun method. That method wait for there
to be any run records with the given status. If the program has
already run, that means it will not wait at all.

Added new startAndWaitForRun methods that will wait for one
additional run of the given status. This method basically
implements the fix that we have had to add to all of these types
of tests, where the number of run records is counted before the
run is started, and it waits for one more run record.